requirements.txt
streamlit==1.50.0
chromadb==1.2.0
numpy==2.3.2
sentence-transformers==5.1.0
python-dotenv==1.1.1
nltk==3.9.2
google-generativeai==0.8.5
mammoth==1.11.0
